Add RoomCreation.ts
